refactor: fix indexing
