+++ src/features/foods/components/Form/index.ts
@@ -1,3 +1,3 @@
+export { AddDrawerContent } from './AddDrawerContent';
 export { CategorySelect } from './CategorySelect';
 export { EditDrawerContent } from './EditDrawerContent';
-export { EditDrawerBody } from './EditDrawerContent/EditDrawerBody';
